Finalized the Updates to the EventDashboard.jsx file. The App.jsx file now uses the EventDashboardWrapper.jsx versus the EventDashboard.jsx.
