+++ frontend/src/components/Events/EventDashboard.jsx
@@ -2,8 +2,6 @@ import React, { useEffect, useState } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
 import { Box, Typography, Card, CardContent, Grid, CircularProgress, Chip, Button, Alert, Table, TableBody, TableCell, TableContainer, TablePagination, TableHead, TableRow, Paper, IconButton, LinearProgress, Drawer, CardHeader, Switch, FormControlLabel, Accordion, AccordionSummary, AccordionDetails, Tabs, Tab, InputAdornment, FormControl, InputLabel, Select, MenuItem, TextField, Autocomplete, Tooltip, Dialog, DialogTitle, DialogContent } from '@mui/material';
 import { useTheme } from '@mui/material/styles';
-import { CheckCircleOutline as CheckCircleIcon, Person as PersonIcon, Groups as GroupsIcon, Assessment as AssessmentIcon, Event as EventIcon, Home as HomeIcon, Menu as MenuIcon, Upload as UploadIcon, PersonAdd as PersonAddIcon, Visibility as VisibilityIcon, VisibilityOff as VisibilityOffIcon, PeopleAlt as PeopleAltIcon, HourglassEmpty as HourglassEmptyIcon, ExpandMore as ExpandMoreIcon, ExpandLess as ExpandLessIcon, CardGiftcard as GiftIcon, Search as SearchIcon, Inventory as InventoryIcon, Save as SaveIcon, Cancel as CancelIcon, Edit as EditIcon, Info as InfoIcon, Refresh as RefreshIcon } from '@mui/icons-material';
-import { BarChart, Bar, XAxis, YAxis, CartesianGrid, Tooltip as RechartsTooltip, Legend, ResponsiveContainer, PieChart, Pie, Cell, LineChart, Line } from 'recharts';
 import api, { fetchInventory } from '../../services/api';
 import { useAuth } from '../../contexts/AuthContext';
 import { usePermissions } from '../../hooks/usePermissions';
@@ -15,392 +13,11 @@ import GuestCheckIn from '../guests/GuestCheckIn';
 import BasicAnalytics from '../dashboard/BasicAnalytics';
 import { getEvent } from '../../services/events';
 import { getEventActivityFeed } from '../../services/api';
-import MainNavigation from '../layout/MainNavigation';
 import ManageSection from './ManageSection';
 import EventHeader from '../events/EventHeader';
-import StandaloneGiftTracker from '../Gifts/StandaloneGiftTracker';
-import EventGiftDashboard from '../Gifts/EventGiftDashboard';
 import GuestTable from '../guests/GuestTable';
+import AdvancedView from './AdvancedView/AdvancedView';
 
-
-const GiftStyleBreakdownTable = () => {
-  const mockGiftData = [
-    { type: 'Tote Bag', style: 'Red', quantity: 18, status: 'Fulfilled' },
-    { type: 'Water Bottle', style: 'Matte Black', quantity: 12, status: 'Pending' }
-  ];
-  const getStatusColor = (status) => {
-    switch (status) {
-      case 'Fulfilled': return 'success';
-      case 'Pending': return 'warning';
-      case 'Shipped': return 'info';
-      default: return 'default';
-    }
-  };
-  return (
-    <Card sx={{ mb: 4 }}>
-      <CardContent>
-        <Typography variant="h6" fontWeight={600} gutterBottom>
-          Gift Style Breakdown
-        </Typography>
-        <TableContainer component={Paper} variant="outlined">
-          <Table>
-            <TableHead>
-              <TableRow sx={{ bgcolor: 'grey.50' }}>
-                <TableCell sx={{ fontWeight: 600 }}>Gift Type</TableCell>
-                <TableCell sx={{ fontWeight: 600 }}>Style</TableCell>
-                <TableCell align="right" sx={{ fontWeight: 600 }}>Quantity Selected</TableCell>
-                <TableCell sx={{ fontWeight: 600 }}>Status</TableCell>
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {mockGiftData.map((item, index) => (
-                <TableRow key={index} hover>
-                  <TableCell>
-                    <Typography variant="body2" fontWeight={500}>{item.type}</Typography>
-                  </TableCell>
-                  <TableCell>
-                    <Typography variant="body2">{item.style}</Typography>
-                  </TableCell>
-                  <TableCell align="right">
-                    <Typography variant="body2" fontWeight={600} color="primary.main">{item.quantity}</Typography>
-                  </TableCell>
-                  <TableCell>
-                    <Chip
-                      label={item.status}
-                      color={getStatusColor(item.status)}
-                      size="small"
-                      sx={{ borderRadius: 1 }}
-                      onClick={undefined}
-                    />
-                  </TableCell>
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
-        </TableContainer>
-      </CardContent>
-    </Card>
-  );
-};
-
-// Fulfillment & Inventory Table using real inventory data
-const FulfillmentInventoryTable = ({ inventory = [] }) => {
-  return (
-    <Card sx={{ mb: 4 }}>
-      <CardContent>
-        <Typography variant="h6" fontWeight={600} gutterBottom>
-          Fulfillment & Inventory
-        </Typography>
-        <TableContainer component={Paper} variant="outlined">
-          <Table>
-            <TableHead>
-              <TableRow sx={{ bgcolor: 'grey.50' }}>
-                <TableCell sx={{ fontWeight: 600 }}>Gift ID</TableCell>
-                <TableCell sx={{ fontWeight: 600 }}>Gift Type</TableCell>
-                <TableCell sx={{ fontWeight: 600 }}>Style</TableCell>
-                <TableCell align="right" sx={{ fontWeight: 600 }}>Qty Warehouse</TableCell>
-                <TableCell align="right" sx={{ fontWeight: 600 }}>Qty On Site</TableCell>
-                <TableCell align="right" sx={{ fontWeight: 600 }}>Current Inventory</TableCell>
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {inventory.map((item) => (
-                <TableRow key={item._id} hover>
-                  <TableCell>{item.sku || item._id}</TableCell>
-                  <TableCell>{item.type}</TableCell>
-                  <TableCell>{item.style}</TableCell>
-                  <TableCell align="right">{item.qtyWarehouse}</TableCell>
-                  <TableCell align="right">{item.qtyOnSite}</TableCell>
-                  <TableCell align="right">{item.currentInventory}</TableCell>
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
-        </TableContainer>
-      </CardContent>
-    </Card>
-  );
-};
-
-// Child Event Summary using real data
-const ChildEventSummary = ({ secondaryEvents = [], guests = [], inventory = [] }) => {
-  // For each child event, count guests and gifts
-  const summary = secondaryEvents.map(child => {
-    const childGuests = guests.filter(g => g.eventId === child._id);
-    // Sum gifts for this child event
-    const childInventory = inventory.filter(i => i.eventId === child._id);
-    const giftCount = childInventory.reduce((sum, i) => sum + (i.currentInventory || 0), 0);
-    return {
-      eventName: child.eventName,
-      giftCount,
-      guestCount: childGuests.length,
-      fulfillmentPercent: childInventory.length > 0 ? Math.round((giftCount / (childInventory.length * 100)) * 100) : 0, // Example
-      status: 'Active', // You can add real status if available
-    };
-  });
-  return (
-    <Card sx={{ mb: 4 }}>
-      <CardContent>
-        <Typography variant="h6" fontWeight={600} gutterBottom>
-          Child Event Summary
-        </Typography>
-        <TableContainer component={Paper} variant="outlined">
-          <Table>
-            <TableHead>
-              <TableRow sx={{ bgcolor: 'grey.50' }}>
-                <TableCell sx={{ fontWeight: 600 }}>Event Name</TableCell>
-                <TableCell align="right" sx={{ fontWeight: 600 }}>Gift Count</TableCell>
-                <TableCell align="right" sx={{ fontWeight: 600 }}>Guest Count</TableCell>
-                <TableCell align="right" sx={{ fontWeight: 600 }}>Fulfillment %</TableCell>
-                <TableCell sx={{ fontWeight: 600 }}>Status</TableCell>
-              </TableRow>
-            </TableHead>
-            <TableBody>
-              {summary.map((row, idx) => (
-                <TableRow key={idx} hover>
-                  <TableCell>
-                    <Typography variant="body2" fontWeight={500}>{row.eventName}</Typography>
-                  </TableCell>
-                  <TableCell align="right">
-                    <Typography variant="body2" fontWeight={600} color="primary.main">{row.giftCount}</Typography>
-                  </TableCell>
-                  <TableCell align="right">
-                    <Typography variant="body2" fontWeight={600}>{row.guestCount}</Typography>
-                  </TableCell>
-                  <TableCell align="right">
-                    <Typography variant="body2" fontWeight={600} color="info.main">{row.fulfillmentPercent}%</Typography>
-                  </TableCell>
-                  <TableCell>
-                    <Chip label={row.status} color={row.status === 'Active' ? 'success' : row.status === 'Completed' ? 'primary' : 'warning'} size="small" sx={{ borderRadius: 1 }} />
-                  </TableCell>
-                </TableRow>
-              ))}
-            </TableBody>
-          </Table>
-        </TableContainer>
-      </CardContent>
-    </Card>
-  );
-};
-
-// Checked-in Guests Chart for AdvancedView
-const CheckedInGuestsChart = ({ guests = [] }) => {
-  const theme = useTheme();
-  const totalGuests = guests.length;
-  const checkedIn = guests.filter(g => g.hasCheckedIn).length;
-  const pending = totalGuests - checkedIn;
-  const pieData = [
-    { name: 'Checked In', value: checkedIn, color: theme.palette.success.main },
-    { name: 'Pending', value: pending, color: theme.palette.warning.main }
-  ];
-  return (
-    <Card sx={{ mb: 4 }}>
-      <CardContent>
-        <Typography variant="h6" fontWeight={600} gutterBottom>
-          Guest Check-In Status
-        </Typography>
-        <ResponsiveContainer width="100%" height={220}>
-          <PieChart>
-            <Pie data={pieData} dataKey="value" nameKey="name" cx="50%" cy="50%" outerRadius={80} label>
-              {pieData.map((entry, idx) => (
-                <Cell key={`cell-${idx}`} fill={entry.color} />
-              ))}
-            </Pie>
-            <Tooltip />
-          </PieChart>
-        </ResponsiveContainer>
-      </CardContent>
-    </Card>
-  );
-};
-
-const AdvancedView = ({ event, guests, secondaryEvents, inventory = [], onInventoryChange }) => {
-  const [activeTab, setActiveTab] = useState(0);
-  const [feedLogs, setFeedLogs] = useState([]);
-  const [feedLoading, setFeedLoading] = useState(false);
-  const [feedType, setFeedType] = useState('');
-  const theme = useTheme();
-  const { eventId } = event || {};
-
-  // Fetch activity feed logs for this event
-  const fetchFeed = async () => {
-    if (!event?._id) return;
-    setFeedLoading(true);
-    try {
-      const res = await getEventActivityFeed(event._id, feedType ? { type: feedType } : {});
-      // Event feed logs loaded
-      setFeedLogs(res.data.logs || []);
-    } catch (err) {
-      console.error('Error loading event feed:', err);
-      setFeedLogs([]);
-    } finally {
-      setFeedLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    fetchFeed();
-  }, [event?._id, feedType]);
-
-  // Refresh activity feed when inventory changes
-  useEffect(() => {
-    if (activeTab === 2) { // Only refresh if on activity feed tab
-      fetchFeed();
-    }
-  }, [inventory, activeTab]); // Refresh when inventory or active tab changes
-
-  // Group inventory by type+style for analytics
-  const groupedByTypeStyle = inventory.reduce((acc, item) => {
-    if (!item.type || !item.style) return acc;
-    const key = `${item.type} - ${item.style}`;
-    if (!acc[key]) acc[key] = 0;
-    acc[key] += item.currentInventory || 0;
-    return acc;
-  }, {});
-
-  // For PieChart: [{ name, value }]
-  const giftSelectionData = Object.entries(groupedByTypeStyle).map(([name, value]) => ({ name, value }));
-
-  // For GiftStyleBreakdownTable: [{ type, style, quantity }]
-  const giftStyleBreakdown = inventory.filter(item => item.type && item.style).map(item => ({
-    type: item.type,
-    style: item.style,
-    quantity: item.currentInventory || 0,
-    status: 'Active', // You can add real status if available
-  }));
-
-  // Use theme palette for pie chart colors
-  const pieColors = [
-    theme.palette.primary.main,
-    theme.palette.secondary.main,
-    theme.palette.success.main,
-    theme.palette.warning.main,
-    theme.palette.info.main,
-    theme.palette.error.main,
-    '#8884d8', '#82ca9d', '#ffc658', '#ff7300', '#00C49F'
-  ];
-
-  // Calculate stats
-  const totalGuests = guests.length;
-  const checkedInGuests = guests.filter(g => g.hasCheckedIn).length;
-  const pendingGuests = totalGuests - checkedInGuests;
-  const checkInPercentage = totalGuests > 0 ? Math.round((checkedInGuests / totalGuests) * 100) : 0;
-
-  // Handler to update notes
-  const handleNoteUpdate = async (inventoryId, newNote) => {
-    try {
-      await api.put(`/inventory/${inventoryId}`, { notes: newNote });
-      if (onInventoryChange) onInventoryChange();
-      // Refresh activity feed after note update
-      fetchFeed();
-    } catch (err) {
-      alert('Failed to update note.');
-    }
-  };
-
-  return (
-    <Box>
-      <Tabs value={activeTab} onChange={(_, v) => setActiveTab(v)} sx={{ borderBottom: 1, borderColor: 'divider', mb: 3 }}>
-        <Tab icon={<GiftIcon />} label="Gift Analytics" iconPosition="start" />
-        <Tab icon={<InventoryIcon />} label="Inventory & Fulfillment" iconPosition="start" />
-        <Tab icon={<InfoIcon />} label="Feed" iconPosition="start" />
-      </Tabs>
-      {activeTab === 0 && (
-        <Box>
-          {/* Gift Style Breakdown Table and PieChart */}
-          <Card sx={{ mb: 4 }}>
-            <CardContent>
-              <Typography variant="h6" fontWeight={600} gutterBottom>
-                Gift Selections by Style
-              </Typography>
-              <TableContainer component={Paper} variant="outlined" sx={{ borderRadius: 2, mb: 3 }}>
-                <Table>
-                  <TableHead>
-                    <TableRow>
-                      <TableCell>Gift Type</TableCell>
-                      <TableCell>Style</TableCell>
-                      <TableCell align="right">Quantity Remaining</TableCell>
-                    </TableRow>
-                  </TableHead>
-                  <TableBody>
-                    {giftStyleBreakdown.map((row, idx) => (
-                      <TableRow key={idx}>
-                        <TableCell>{row.type}</TableCell>
-                        <TableCell>{row.style}</TableCell>
-                        <TableCell align="right">{row.quantity}</TableCell>
-                      </TableRow>
-                    ))}
-                  </TableBody>
-                </Table>
-              </TableContainer>
-              <ResponsiveContainer width="100%" height={300}>
-                <PieChart>
-                  <Pie data={giftSelectionData} dataKey="value" nameKey="name" cx="50%" cy="50%" outerRadius={100} label>
-                    {giftSelectionData.map((entry, idx) => (
-                      <Cell key={`cell-${idx}`} fill={pieColors[idx % pieColors.length]} />
-                    ))}
-                  </Pie>
-                  <RechartsTooltip />
-                </PieChart>
-              </ResponsiveContainer>
-            </CardContent>
-          </Card>
-        </Box>
-      )}
-      {activeTab === 1 && <FulfillmentInventoryTable inventory={inventory} />}
-      {activeTab === 2 && (
-        <Box>
-          <Box sx={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', mb: 2 }}>
-            <Typography variant="h6">Event Activity Feed</Typography>
-            <Tooltip title="Refresh Activity Feed">
-              <IconButton onClick={fetchFeed} disabled={feedLoading}>
-                <RefreshIcon />
-              </IconButton>
-            </Tooltip>
-          </Box>
-        </Box>
-      )}
-    </Box>
-  );
-};
-
-// Event Dashboard Wrapper that fetches inventory and passes it to EventDashboard
-const EventDashboardWrapper = () => {
-  const { eventId } = useParams();
-  const [inventory, setInventory] = useState([]);
-  const [inventoryLoading, setInventoryLoading] = useState(false);
-  const [inventoryError, setInventoryError] = useState('');
-
-  const loadInventory = async () => {
-    if (!eventId) return;
-
-    setInventoryLoading(true);
-    setInventoryError('');
-    try {
-      const response = await fetchInventory(eventId);
-      setInventory(response.data.inventory || []);
-    } catch (error) {
-      console.error('Error fetching inventory:', error);
-      setInventoryError('Failed to fetch inventory data');
-    } finally {
-      setInventoryLoading(false);
-    }
-  };
-
-  useEffect(() => {
-    loadInventory();
-  }, [eventId]);
-
-  return (
-    <EventDashboard
-      eventId={eventId}
-      inventory={inventory}
-      inventoryLoading={inventoryLoading}
-      inventoryError={inventoryError}
-      onInventoryChange={loadInventory}
-    />
-  );
-};
 
 const EventDashboard = ({ eventId, inventory = [], inventoryLoading = false, inventoryError = '', onInventoryChange }) => {
   const navigate = useNavigate();
@@ -648,4 +265,4 @@ const EventDashboard = ({ eventId, inventory = [], inventoryLoading = false, inv
   );
 };
 
-export default EventDashboardWrapper;
+export default EventDashboard;
